Add PodTemplateSpecBuilder
